+++ lib/dubp/user-docs/src/documents/revocation/v10.rs
@@ -193,7 +193,7 @@ impl RevocationDocumentV10 {
 impl Document for RevocationDocumentV10 {
     type PublicKey = PubKey;
 
-    fn version(&self) -> u16 {
+    fn version(&self) -> usize {
         10
     }
