+++ lib/dubp/user-docs/src/documents/transaction.rs
@@ -715,7 +715,7 @@ impl TransactionDocument {
 impl Document for TransactionDocument {
     type PublicKey = PubKey;
 
-    fn version(&self) -> u16 {
+    fn version(&self) -> usize {
         10
     }
